+++ src/constants.rs
@@ -1,9 +1,9 @@
 pub const UPSAMPLE_TARGET_FREQUENCY: u32 = 192_000;
 pub const LOW_PASS_FILTER_SIZE: usize = 48;
 
-pub const MAX_8_BIT: f32 = 127.0;
-pub const MAX_16_BIT: f32 = 32767.0;
-pub const MAX_24_BIT: f32 = 8388607.0;
-pub const MAX_32_BIT: f32 = 2147483647.0;
+pub const MAX_8_BIT: f64 = 127.0;
+pub const MAX_16_BIT: f64 = 32767.0;
+pub const MAX_24_BIT: f64 = 8388607.0;
+pub const MAX_32_BIT: f64 = 2147483647.0;
 
 pub const UNITS: &[&str] = &["B", "KB", "MB", "GB", "TB"];
